API: libcrmcommon: Deprecate mainloop_timer_t API
Deprecate the mainloop_timer_t type itself and all functions that accept
or manipulate mainloop_timer_t objects. External programs should not use
Pacemaker for general-purpose asynchronous programming.
